Move lims_queries and update imports
