+++ ipfx/bin/generate_pipeline_input.py
@@ -3,7 +3,7 @@ import ipfx.qc_feature_evaluator as qcp
 import allensdk.core.json_utilities as ju
 import os.path
 from ipfx.bin.generate_se_input import generate_se_input, parse_args
-import ipfx.bin.lims_queries as lq
+import ipfx.lims_queries as lq
 
 
 def generate_pipeline_input(cell_dir=None,
