+++ ipfx/bin/generate_se_input.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 import allensdk.core.json_utilities as ju
 import os
 from . import make_stimulus_ontology as mso
-from . import lims_queries as lq
+import ipfx.lims_queries as lq
 import argparse
